+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## Unreleased
+## 0.2
+
+- Add tooltip when hovering node containing multiple threads
+- Improve node link rendering
+- Display information when the thread graph is not available
 
 ## 0.1